make the mdx setup work
