+++ src/components/MDX/ErrorDecoder.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import {useEffect, useState} from 'react';
 import {useErrorDecoderParams} from '../ErrorDecoderContext';
 import cn from 'classnames';
