+++ src/components/MDX/ExpandableExample.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 /*
  * Copyright (c) Facebook, Inc. and its affiliates.
  */
